Avoid re-rendering tx list on header btn click
